年収、身長項目の書き込み処理等まで完了 プロフ用画像(img_name2,3)処理は未着手。
掲載用写真テーブル、趣味マスターテーブル、休日マスターテーブル作成
(中間テーブルの使い方おさらいするのでここの実装は少々いただきます。)

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -29,7 +29,7 @@ class CreateUsersTable extends Migration
             $table->string('password');
             $table->tinyInteger('smoke')->default(0);
             $table->tinyInteger('alcohol')->default(0);
-            $table->float('body_height',4,1)->nullable();
+            $table->integer('body_height')->nullable();
             $table->tinyInteger('body_figure')->default(1);
             $table->tinyInteger('education')->nullable();
             $table->tinyInteger('housemate')->nullable();
